moved and renamed mock file
